Rahul/move split secret (#4557)
* moving split secret to frost_utils

* removing unused import

* removing double &

* source code heading
