+++ ironfish-rust/src/transaction/mod.rs
@@ -21,17 +21,7 @@ use crate::{
     OutgoingViewKey, OutputDescription, SpendDescription, ViewKey,
 };
 
-use ironfish_frost::frost;
-use ironfish_frost::frost::Error;
-use rand::{
-    rngs::{OsRng, ThreadRng},
-    thread_rng,
-};
-
-use ironfish_frost::frost::{
-    keys::{IdentifierList, KeyPackage, PublicKeyPackage},
-    Identifier, SigningKey,
-};
+use rand::{rngs::OsRng, thread_rng};
 
 use bellperson::groth16::{verify_proofs_batch, PreparedVerifyingKey};
 use blake2b_simd::Params as Blake2b;
@@ -49,7 +39,6 @@ use ironfish_zkp::{
 };
 
 use std::{
-    collections::HashMap,
     io::{self, Write},
     iter,
     slice::Iter,
@@ -946,45 +935,4 @@ pub fn batch_verify_transactions<'a>(
         &SAPLING.output_verifying_key,
         &SAPLING.mint_verifying_key,
     )
-}
-
-pub struct SecretShareConfig {
-    pub min_signers: u16,
-    pub max_signers: u16,
-    pub secret: Vec<u8>,
-}
-
-pub fn split_secret(
-    config: &SecretShareConfig,
-    identifiers: IdentifierList,
-    rng: &mut ThreadRng,
-) -> Result<(HashMap<Identifier, KeyPackage>, PublicKeyPackage), Error> {
-    let secret_key = SigningKey::deserialize(
-        config
-            .secret
-            .clone()
-            .try_into()
-            .map_err(|_| Error::MalformedSigningKey)?,
-    )?;
-
-    let (shares, pubkeys) = frost::keys::split(
-        &secret_key,
-        config.max_signers,
-        config.min_signers,
-        identifiers,
-        rng,
-    )?;
-
-    for (_k, v) in shares.clone() {
-        frost::keys::KeyPackage::try_from(v)?;
-    }
-
-    let mut key_packages: HashMap<_, _> = HashMap::new();
-
-    for (identifier, secret_share) in shares {
-        let key_package = frost::keys::KeyPackage::try_from(secret_share.clone()).unwrap();
-        key_packages.insert(identifier, key_package);
-    }
-
-    Ok((key_packages, pubkeys))
 }
